### 2017-10-01 #420 @whedon
[PRE REVIEW]: chloroExtractor: extraction and assembly of the chloroplast genome from whole genome shotgun data
**Submitting author:** @greatfireball (<a href="http://orcid.org/0000-0003-4166-5423">Frank Förster</a>)
**Repository:** <a href="https://github.com/chloroExtractorTeam/chloroExtractor" target ="_blank">https://github.com/chloroExtractorTeam/chloroExtractor</a>
**Version:** v1.0.0
