### 2017-11-13 #420 @greatfireball
Morning, @tracykteal, @whedon, and @arfon after six weeks without anything happen at this issue, I just wanted to ask if we need to provide something I missed or what we can to provide to speed things up.

Thanks in advance!
