### 2017-11-14 #420 @arfon
> Morning, @tracykteal, @whedon, and @arfon after six weeks without anything happen at this issue, I just wanted to ask if we need to provide something I missed or what we can to provide to speed things up.

Yep, looks like we need to find an alternative editor for this submission.

@pjotrp @biorelated @mgymrek - would one of you be willing to edit this submission for JOSS?
