### 2017-11-17 #464 @pjotrp
Thanks @jdeligt for being willing to review this submission! Make sure to accept the invite to the 'openjournals/joss-reviews' repository to be able to tick above boxes and read the reviewer instructions http://joss.theoj.org/about. @greatfireball it is very helpful if you can go through above checklist first and make sure the important points are implemented. Please reply here if you think it all hunky dory!
