### 2017-11-17 chloroExtractorTeam/chloroExtractor#80 @jdeligt
Minor edit to add quotes in paper.md
While I'm going through the works on the software you could already address this very minor point. Please quote multiple affiliations as specified in the guidelines.

authors:
 - name: Arfon M Smith
   orcid: 0000-0000-0000-1234
   affiliation: "1, 2" # (Multiple affiliations must be quoted)
