### 2017-11-17 #464 @jdeligt
I'll wait with further review until @greatfireball has gone through the checklist.
Found one small issue in the paper.md that has been raised on their repo:
https://github.com/chloroExtractorTeam/chloroExtractor/issues/80
