### 2017-11-17 chloroExtractorTeam/chloroExtractor#81 @jdeligt
Example in docker does not work
When I run the example usage:

docker pull chloroextractorteam/chloroextractor # ensure the latest version from docker hub
docker run -v /tmp/chloroExtractor-testrun:/data --rm chloroextractorteam/chloroextractor -1 SRR5216995_1M_1.fastq -2 SRR5216995_1M_2.fastq

It returns an exit state 9:

>All dependencies fulfilled!
>';
>[17-11-17 13:13:23] [PipeWrap] Running 'jf0': jellyfish count -t 2 -m 31 -s 500M -C -o jf0.jf >/data/SRR5216995_1M_1.fastq /data/SRR5216995_1M_2.fastq
>[17-11-17 13:13:36] [PipeWrap] jf0 exited:9
>[17-11-17 13:13:36] [PipeWrap]  at /opt/chloroExtractor/bin/../lib/PipeWrap.pm line 153.
