### 2017-11-17 #464 @jdeligt
And please check the docker image and example usage, currently it fails:
https://github.com/chloroExtractorTeam/chloroExtractor/issues/81
