### 2017-11-17 #464 @jdeligt
As well as the binary download:
> [17-11-17 15:57:06] [fcg] Unable to open file 'ass/assembly_graph.fastg'
> [17-11-17 15:57:06] [PipeWrap] fcg exited:256
> [17-11-17 15:57:06] [PipeWrap]  at /Users/jdeligt/code/chloroExtractor/bin/../lib/PipeWrap.pm line 153
