### 2017-11-22 chloroExtractorTeam/chloroExtractor#81 @iimog
Thanks for raising this issue. This is a documentation problem. I assume this only works when there is a folder /tmp/chloroExtractor-testrun with the two SRR fastq files on the host machine, @greatfireball can you please verify and fix the documentation?
Otherwise I can have a closer look when I'm back from South Africa in early December.
