### 2017-11-24 chloroExtractorTeam/chloroExtractor#81 @jdeligt
I've bound the data folder in the docker to the right local directory:
docker run -v /Users/jdeligt/data/chloroExtractor/:/data

creating the  /tmp/chloroExtractor-testrun on my machine does help but raises a new issue:
>[17-11-24 07:36:46] [PipeWrap] Running 'fcg': /opt/chloroExtractor/bin/find_cyclic_graph.pl --in ass/assembly_graph.fastg --blastdb /opt/chloroExtractor/bin//../data/cds.nr98.fa --out fcg.fa
>[17-11-24 07:36:46] [fcg] Unable to open file 'ass/assembly_graph.fastg'
>[17-11-24 07:36:46] [PipeWrap] fcg exited:256
>[17-11-24 07:36:46] [PipeWrap]  at /opt/chloroExtractor/bin/../lib/PipeWrap.pm line 153.

Drop me a line when the docker and source documentations are in working order so I can continue the review.
