### 2017-12-05 4f6823a
Add quotes for multiple affiliations
Fix #80
