### 2017-12-05 chloroExtractorTeam/chloroExtractor#81 @DanielAmsel
Hi *,
I ran into the same problem as @jdeligt .
So I assume we both use a macOS system?
I managed to overcome the issue by simply increasing the memory, docker wants to allocate.
In my case, I allowed 5GB (but I think anything larger than the default 2GB would fit).
( https://stackoverflow.com/a/39720010 )

My settings were:
MacBook Pro 13
macOS High Sierra 10.13.1
2.7GHz Intel Core I5
8GB RAM (1867Mhz DDR3)

I took the commands from
https://github.com/chloroExtractorTeam/chloroExtractor#example

```bash
mkdir -p /tmp/chloroExtractor-testrun
cd /tmp/chloroExtractor-testrun
wget 'https://zenodo.org/record/884449/files/SRR5216995_1M.tar.bz2' -O - | tar xjf -
docker pull chloroextractorteam/chloroextractor
docker run -v /tmp/chloroExtractor-testrun:/data --rm chloroextractorteam/chloroextractor -1 SRR5216995_1M_1.fastq -2 SRR5216995_1M_2.fastq
```
Best,
Daniel
