### 2017-12-05 chloroExtractorTeam/chloroExtractor#81 @jdeligt
This does indeed seem to solve the issue.
Please update the documentation to reflect this and the requirement for the directory.

The issue with the source code version is not related to this, do you want me to open a separate issue for this?
