### 2017-12-05 #464 @jdeligt
Dependencies do not contain version numbers
