### 2017-12-06 chloroExtractorTeam/chloroExtractor#83 @iimog
Requirements do not contain version numbers
The list of requirements in the README does not contain version numbers as noted in https://github.com/openjournals/joss-reviews/issues/464
I'll list the version numbers from the current docker container as versions which are known to work. We should probably note that other versions might work as well but are not guaranteed to. We should also discuss some way to keep this listing up to date.
