### 2017-12-06 chloroExtractorTeam/chloroExtractor#83 @iimog
Required Software:
 - [ ] Jellyfish
 - [ ] Spades
 - [ ] bowtie2
 - [ ] NCBI-Blast+
 - [ ] Samtools
 - [ ] Bedtools
 - [ ] GNU R
 - [ ] Ghostscript
 - [ ] Python
 - [ ] Perl

Required Perl modules:
 - [ ] Moose
 - [ ] Log::Log4Perl
 - [ ] Term::ProgressBar
 - [ ] Graph
 - [ ] IPC::Run
 - [ ] File::Which
