### 2017-12-06 4ea742c
Add version numbers to requirements. Fix #83
