### 2017-12-06 5d87a72
Add docu for memory requirement and adjustment for Mac OS X. Related to #81
