### 2017-12-06 af7b591
Add docu for data directory in example. Related to #81
