### 2017-12-06 cb1c517
Make usage of demo data folder more generic and consistent.
