### 2017-12-06 c010a63
Use consistent form for shell variables (with braces)
