### 2017-12-07 chloroExtractorTeam/chloroExtractor#81 @iimog
The documentation is now up to date with version numbers for requirements, information about docker memory requirements and a more consistent description of the demo directory.

Regarding the issue with the source code version, I'm not sure if I can reproduce it. If you don't mind @jdeligt it would be nice to open a separate issue for that.
