### 2017-12-07 chloroExtractorTeam/chloroExtractor#81 @jdeligt
Will do. Hopefully tomorrow. I will add as much info as possible to
reproduce.

On Thu, 7 Dec 2017, 09:07 Markus J. Ankenbrand, <notifications@github.com>
wrote:

> The documentation is now up to date with version numbers for requirements,
> information about docker memory requirements and a more consistent
> description of the demo directory.
>
> Regarding the issue with the source code version, I'm not sure if I can
> reproduce it. If you don't mind @jdeligt <https://github.com/jdeligt> it
> would be nice to open a separate issue for that.
>
