### 2017-12-18 #464 @greatfireball
Hey @pjotrp and @jdeligt sorry for the delay. Think we could close those issues above by https://github.com/chloroExtractorTeam/chloroExtractor/issues/83 and https://github.com/chloroExtractorTeam/chloroExtractor/issues/81. @jdeligt mentioned another issue, but we need some more information to reproduce his finding. Nevertheless, the his major concerns should be solved. Please let us know, how we can help you to continue the review process. Thanks again for your time, your help, and your review.

Best,
Frank
