### 2018-01-04 267250c
Added citations for Bowtie2, Blast+, samtools, bedtools and R
