+++ paper/paper.bib
@@ -176,3 +176,66 @@ and Zhang, Xuegong},
   Url                      = { + http://dx.doi.org/10.1093/bioinformatics/bth352}
 }
 
+@Article{bowtie2,
+ Title                     = {Fast gapped-read alignment with Bowtie 2},
+ Author                    = {Langmead B, Salzberg S.},
+ Journal                   = {Nature Methods},
+ Year                      = {2012},
+ Page                      = {357-359},
+ Volume                    = {9},
+ 
+ Doi                       = {10.1038/nmeth.1923},
+ Url                       = {http://dx.doi.org/10.1038/nmeth.1923}
+}
+
+@Article{blast+,
+ Title                     = {BLAST+: architecture and applications},
+ Author                    = {Camacho C, Coulouris G, Avagyan V, Ma N, Papadopoulos J, Bealer K, Madden TL.},
+ Journal                   = {BMC Bioinformatics},
+ Year                      = {2009},
+ Number                    = {1},
+ Month                     = {Dec},
+ Page                      = {421},
+ Volume                    = {10},
+
+ Doi                       = {10.1186/1471-2105-10-421},
+ Url                       = {https://doi.org/10.1186/1471-2105-10-421}
+}
+
+@Article{Samtools,
+ Title                     = {The Sequence Alignment/Map format and SAMtools},
+ Autor                     = {Li H, Handsaker B, Wysoker A, Fennell T, Ruan J, Homer N, Marth G, Abecasis G, Durbin R. and 1000 Genome Project Data Processing Subgroup},
+ Journal                   = {Bioinformatics},
+ Year                      = {2009},
+ Number                    = {1},
+ Month                     = {Aug},
+ Page                      = {2078-2079},
+ Volume                    = {25},
+
+ Doi                       = {10.1093/bioinformatics/btp352},
+ Url                       = {https://doi.org/10.1093/bioinformatics/btp352}
+}
+
+@Article{Bedtools,
+ Title                     = {BEDTools: a flexible suite of utilities for comparing genomic features},
+ Author                    = {Aaron R. Quinlan, Ira M. Hall},
+ Journal                   = {Bioinformatics},
+ Year                      = {2010},
+ Number                    = {1},
+ Month                     = {Mar},
+ Page                      = {841-842},
+ Volume                    = {26},
+
+ Doi                       = {10.1093/bioinformatics/btq033},
+ Url                       = {https://doi.org/10.1093/bioinformatics/btq033}
+}
+
+@Article{GnuR,
+ Title        = {R: A Language and Environment for Statistical Computing},
+ Author       = {{R Core Team}},
+ Organization = {R Foundation for Statistical Computing},
+ Address      = {Vienna, Austria},
+ Year         = {2008},
+ Note         = {note = {{ISBN} 3-900051-07-0},
+ Url          = {https://www.R-project.org}
+}
