### 2018-01-04 418c697
Added Ghostsccript, Phyton, Perl, Perl_moduls(Moose, Log::Log4Perl, File::Which, IPC::Run, Graph, Term::Progress) as inline Links in paper.md
