### 2018-01-04 57ee054
corrected new added Authors to Format:'Surrname, Name and' like older citations
