+++ paper/paper.bib
@@ -178,7 +178,7 @@ and Zhang, Xuegong},
 
 @Article{bowtie2,
  Title                     = {Fast gapped-read alignment with Bowtie 2},
- Author                    = {Langmead B, Salzberg S.},
+ Author                    = {Langmead, Ben and Salzberg, Steven L.},
  Journal                   = {Nature Methods},
  Year                      = {2012},
  Page                      = {357-359},
@@ -190,7 +190,7 @@ and Zhang, Xuegong},
 
 @Article{blast+,
  Title                     = {BLAST+: architecture and applications},
- Author                    = {Camacho C, Coulouris G, Avagyan V, Ma N, Papadopoulos J, Bealer K, Madden TL.},
+ Author                    = {Camacho, Christiam and Coulouris, George and Avagyan, Vahram and Ma, Ning and Papadopoulos, Jason and Bealer, Kevin and Madden, Thomas L.},
  Journal                   = {BMC Bioinformatics},
  Year                      = {2009},
  Number                    = {1},
@@ -204,7 +204,7 @@ and Zhang, Xuegong},
 
 @Article{Samtools,
  Title                     = {The Sequence Alignment/Map format and SAMtools},
- Autor                     = {Li H, Handsaker B, Wysoker A, Fennell T, Ruan J, Homer N, Marth G, Abecasis G, Durbin R. and 1000 Genome Project Data Processing Subgroup},
+ Autor                     = {Li, Heng and Handsaker, Bob and Wysoker, Alec and Fennell, Tim and Ruan, Jue and Homer, Nils and Marth, Gabor and Abecasis, Goncalo and Durbin, Richard and 1000 Genome Project Data Processing Subgroup},
  Journal                   = {Bioinformatics},
  Year                      = {2009},
  Number                    = {1},
@@ -218,7 +218,7 @@ and Zhang, Xuegong},
 
 @Article{Bedtools,
  Title                     = {BEDTools: a flexible suite of utilities for comparing genomic features},
- Author                    = {Aaron R. Quinlan, Ira M. Hall},
+ Author                    = {Quinland, Aaron R. and Hall, Ira M.},
  Journal                   = {Bioinformatics},
  Year                      = {2010},
  Number                    = {1},
