### 2018-01-05 b550a8b
Add links to other used tools and modules
