+++ paper/paper.md
@@ -70,22 +70,12 @@ In the future, we plan to use our chloroExtractor to screen NCBI's Sequence Read
 
 ![Schematic workflow of chloroExtractor.](workflow.png){ width=70% }
 
+In addition to the components cited above the chloroExtractor uses [Ghostscript](https://www.ghostscript.com/), [Phyton](https://www.python.org/), and [Perl](https://www.perl.org/).
+Further the following Perl modules are used: [Moose](http://search.cpan.org/~ether/Moose-2.2009/lib/Moose.pm), [Log::Log4Perl](http://search.cpan.org/~mschilli/Log-Log4perl-1.49/lib/Log/Log4perl.pm), [Graph](http://search.cpan.org/dist/Graph/lib/Graph.pod), [Term::ProgressBar](http://search.cpan.org/~manwar/Term-ProgressBar-2.21/lib/Term/ProgressBar.pm), [IPC::Run](http://search.cpan.org/~toddr/IPC-Run-0.96/lib/IPC/Run.pm), and [File::Which](http://search.cpan.org/~plicease/File-Which-1.22/lib/File/Which.pm).
+
 # Acknowledgements
 
 MJA was supported by a grant of the German Excellence Initiative to the Graduate School of Life Sciences, University of Würzburg.
 
 # References
-
-# Citations
-[Ghostscript](https://www.ghostscript.com/)
-[Phyton](https://www.python.org/)
-[Perl](https://www.perl.org/)
-
-Perl Modules
-[Moose](http://search.cpan.org/~ether/Moose-2.2009/lib/Moose.pm)
-[Log::Log4Perl](http://search.cpan.org/~mschilli/Log-Log4perl-1.49/lib/Log/Log4perl.pm)
-[Graph](http://search.cpan.org/dist/Graph/lib/Graph.pod)
-[Term::ProgressBar](http://search.cpan.org/~manwar/Term-ProgressBar-2.21/lib/Term/ProgressBar.pm)
-[IPC::Run](http://search.cpan.org/~toddr/IPC-Run-0.96/lib/IPC/Run.pm)
-[File::Which](http://search.cpan.org/~plicease/File-Which-1.22/lib/File/Which.pm)
 
