+++ paper/paper.bib
@@ -188,7 +188,7 @@ and Zhang, Xuegong},
  Url                       = {http://dx.doi.org/10.1038/nmeth.1923}
 }
 
-@Article{blast+,
+@Article{blastPlus,
  Title                     = {BLAST+: architecture and applications},
  Author                    = {Camacho, Christiam and Coulouris, George and Avagyan, Vahram and Ma, Ning and Papadopoulos, Jason and Bealer, Kevin and Madden, Thomas L.},
  Journal                   = {BMC Bioinformatics},
@@ -236,6 +236,5 @@ and Zhang, Xuegong},
  Organization = {R Foundation for Statistical Computing},
  Address      = {Vienna, Austria},
  Year         = {2008},
- Note         = {note = {{ISBN} 3-900051-07-0},
  Url          = {https://www.R-project.org}
 }
