### 2018-01-06 #464 @greatfireball
Happy new year to both of you @pjotrp and @jdeligt, hope you had a wonderful holiday season. We updated our paper.md and added some more citations. We are still waiting for the more detailed description of the issue @jdeligt found. Please let us know if we can help you finishing the review process.
