### 2018-01-10 #464 @jdeligt
Thanks for the reminder and updating the documentation and contribution sections. The latest & greatest binary is running now, if it produces errors I will raise a well documented issue, if not I will tick the last box and the review will be concluded.
