### 2018-01-10 chloroExtractorTeam/chloroExtractor#88 @jdeligt
Use of relative paths (seen with spades)
There seem to be some issues in the way you use spades (3.11.0) with relative paths:

running this:
`assemble_spades.pl -1 kfr2_1.fq -2 kfr2_2.fq -o ass --threads 2`

results in this:
` /usr/local/bin/spades.py	--only-assembler	-1	/private/tmp/chloroExtractor-testrun/ptx/kfr2_1.fq	-2	/private/tmp/chloroExtractor-testrun/ptx/kfr2_2.fq	-o	/private/tmp/chloroExtractor-testrun/ptx/ass`

while the data/working directory is this:
`/tmp/chloroExtractor-testrun/ptx`

specifying full paths resolves this:
```
assemble_spades.pl -1 /tmp/chloroExtractor-testrun/ptx/kfr2_1.fq -2 /tmp/chloroExtractor-testrun/ptx/kfr2_2.fq -o /tmp/chloroExtractor-testrun/ptx/ass
->
/usr/local/bin/spades.py	--only-assembler	-1	/tmp/chloroExtractor-testrun/ptx/kfr2_1.fq	-2	/tmp/chloroExtractor-testrun/ptx/kfr2_2.fq	-o	/tmp/chloroExtractor-testrun/ptx/ass
```
