### 2018-01-10 #464 @jdeligt
Apart from a minor issue with relative paths https://github.com/chloroExtractorTeam/chloroExtractor/issues/88 no further problems were encountered and this issue can be worked around by the average user. Seen how the previous issues were handled I am confident that this will be resolved and should not impede publication.

As such I'm closing this review and would recommend publication.
