### 2018-01-11 chloroExtractorTeam/chloroExtractor#88 @greatfireball
Thx for the report @jdeligt

Did you call `assemble_spades.pl` directly or using our chloroExtractor pipeline?

I guess it is a Mac specific issue. I found the following [link](https://apple.stackexchange.com/questions/1043/why-is-tmp-a-symlink-to-private-tmp) at stackexchange. The explain that `/tmp/` is a symlink to `/private/tmp/`. In that case it might be correct, that the local files get expanded paths like `/private/tmp/...`. If you are providing absolute path, the expansion into absolute paths is not required and therefore the link is not resolved into its `/private/tmp/` equivalent.

Nevertheless, I will try to find a Mac to verify that guess.
