### 2018-01-16 #464 @greatfireball
Of course I can :) I did a copy of the current status and pushed it to Zenodo. Its DOI is [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1147434.svg)](https://doi.org/10.5281/zenodo.1147434) and the current version number was bumped to v1.0.1

@jdeligt, @pjotrp, and @arfon thanks for your efforts!
