### 2018-01-16 chloroExtractorTeam/chloroExtractor#81 @greatfireball
As the issue was solved for @jdeligt I will close that issue
