### 2018-01-16 #464 @greatfireball
I would like to add the guy who did solve the Mac-issues with our container to the acknowledge section of the paper. Would be nice, if that would be possible. I will then create another version for zenodo. So sry @arfon for the delay. Just give me an hour.
