### 2018-01-16 #464 @greatfireball
The new DOI is [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1148955.svg)](https://doi.org/10.5281/zenodo.1148955) and the current version is 1.0.2

Thanks again!
