### 2018-01-16 #464 @greatfireball
Yeah... Version v1.0.2 contains @DanielAmsel in the acknowledgements section
